Create the_number_is_repeated_as_many_times_as_it_is.py
Write a program that displays part of the sequence 1 2 2 3 3 3 4 4 4 4 5 5 5 5 5 ... (the number is repeated as many times as it is). A non-negative integer n is passed to the program - the program should display the number of sequence elements. The output is an expected sequence of numbers written with a single line space.

For example, if n = 7, then the program should output 1 2 2 3 3 3 4.