Create changing_matrix_elements.py
Write a program whose input is a rectangular matrix as a sequence of strings ending in a string containing only the string "end" (without quotes)

The program should output a matrix of the same size, in which each element in position i, j is equal to the sum of the elements of the first matrix in positions (i-1, j), (i + 1, j), (i, j-1), (i j + 1). In the extreme symbols, the neighboring element is located on the opposite side of the matrix.

In the case of a single row/column, the element itself is a neighbor in the corresponding direction.